// src/utils/web3/types.ts
export interface JBSplit {
	preferClaimed: boolean;
	preferAddToBalance: boolean;
	percent: BigNumber | number | string;
	projectId: BigNumber | number;
	beneficiary: string;
	lockedUntil: BigNumber | number;
	allocator: string;
}
// ...
export interface JBProjectMetadata {
	content: string;
	domain: number;
}
// ...
export interface JBFundingCycleData {
	duration: number | BigNumber;
	weight: number | BigNumber;
	discountRate: number | BigNumber;
	ballot: string;
}

export interface JBGlobalFundingCycleMetadata {
	allowSetTerminals: boolean;
	allowSetController: boolean;
	pauseTransfers: boolean;
}

export interface JBFundingCycleMetadata {
	global: JBGlobalFundingCycleMetadata;
	reservedRate: number | BigNumber;
	redemptionRate: number | BigNumber;
	ballotRedemptionRate: number | BigNumber;
	pausePay: boolean;
	pauseDistributions: boolean;
	pauseRedeem: boolean;
	pauseBurn: boolean;
	allowMinting: boolean;
	allowTerminalMigration: boolean;
	allowControllerMigration: boolean;
	holdFees: boolean;
	preferClaimedTokenOverride: boolean;
	useTotalOverflowForRedemptions: boolean;
	useDataSourceForPay: boolean;
	useDataSourceForRedeem: boolean;
	dataSource: string;
	metadata: number;
}

export interface JBGroupedSplits {
	group: number;
	splits: JBSplit[];
}

export interface JBFundAccessConstraints {
	terminal: string;
	token: string;
	distributionLimit: number | BigNumber;
	distributionLimitCurrency: number | BigNumber;
	overflowAllowance: number | BigNumber;
	overflowAllowanceCurrency: number | BigNumber;
}
// ...
export const JBTOKENS_ETH = '0x000000000000000000000000000000000000EEEe';